ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.8.2
pymupdf==1.23.8
numpy==1.26.3
transformers==4.36.2
torch>=2.2.0
sentence-transformers==2.3.1
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
stable-baselines3==2.2.1
gymnasium==0.29.1
shimmy>=0.2.1
edge-tts==6.1.9
openai-whisper==20231117
aiofiles==23.2.1
requests==2.31.0
scikit-learn==1.4.0
joblib==1.3.2